Move Toolbar to different folder
