+++ frontend/src/pages/BoardPage.tsx
@@ -1,5 +1,5 @@
 import BoardContainer from "@/components/board/BoardContainer"
-import LoggedInContainer from "@/components/LoggedInContainer"
+import LoggedInContainer from "@/components/general/LoggedInContainer"
 
 const ChartsContainer = () => {
   return <LoggedInContainer>{(props) => <BoardContainer board={props.board} />}</LoggedInContainer>
